EnergyStatistics.jl

In statistics distance correlation or distance covariance is a measure of dependence between two paired random vectors. The population distance correlation coefficient is zero if and only if the random vectors are independent. Thus, distance correlation measures both linear and nonlinear association between two random vectors. This is in contrast to Pearson's correlation, which can only detect linear association between two random variables. See here for references and more details.

Installation

This package can be installed using the Julia package manager. From the Julia REPL, type ] to enter the Pkg REPL mode and run

pkg> add EnergyStatistics

General Usage

Given two random vectors x and y the distance correlation dcor can simply computed:

using EnergyStatistics

x = collect(-1:0.01:1)
y = map(x -> x^4 - x^2, x)

dcor(x, y) ≈ 0.374204050

These two vectors are clearly associated. However, their (Pearson) correlation coefficient vanishes suggesting that they are independent. The finite distance correlation dcor reveals their non-linear association.

Function to compute the distance covariance dcov and distance variance dvar are also supplied.

Advanced Usage

The computation of the 'distance matrices' is computationally expansive. In cases where you want to compute several distance correlation keeping one of vectors fixed you may consider storing the distance matrices for future use.

For example

A = dcenter!(DistanceMatrix(x))
B = dcenter!(DistanceMatrix(y))
C = dcenter!(DistanceMatrix(z))

dcor(A, B)
dcor(A, C)

will run much faster than

dcor(x, y)
dcor(x, z)

since the distance matrix for the vector x has to be compute only once.

You can also construct distance matrices using other distance measures than the (default) abs.

AA = dcenter!(DistanceMatrix(Float64, x, abs2))

In place double centering dcenter! and U-centering ucenter! functions are available.

References

See the wikipedia page for references and more details.

Functions

EnergyStatistics.dcor — Method

dcor(x::AbstractVector{T}, y::AbstractVector{T}) where T <: Real

Computes the distance correlation of samples x and y.

using EnergyStatistics
x = collect(-1:0.01:1)
y = @. x^4 - x^2
dcor(x, y)

# output

0.3742040504583154

EnergyStatistics.dcov — Method

dcov(x::AbstractVector{T}, y::AbstractVector{T}) where T <: Real

Computes the distance covariance of samples x and y.

EnergyStatistics.dvar — Method

dvar(x::AbstractVector{T}) where T <: Real

Computes the distance variance of a sample x.

EnergyStatistics.DistanceMatrix — Method

DistanceMatrix(x::AbstractVector{T}, dist = abs) where {T}

Computes the matrix of pairwise distance of x. The distance measure dist is abs as default.

using EnergyStatistics
x = [1.0, 2.0]
EnergyStatistics.DistanceMatrix(x)

# output

2×2 Array{Float64,2}:
 0.0  1.0
 1.0  0.0

EnergyStatistics.dcenter! — Method

dcenter!(A::DistanceMatrix{T}) where {T <: Real}

Computes the double centered matrix of A in place.

EnergyStatistics.ucenter! — Method

ucenter!(A::DistanceMatrix{T}) where {T <: Real}

Computes the u-centered matrix of A in place.

EnergyStatistics.dcor — Method

dcor(A::DistanceMatrix{T}, B::DistanceMatrix{T}, dvarA::T = dvar(A), dvarB::T = dvar(B))

Computes the distance correlation of two DistanceMatrices A and B.

EnergyStatistics.dcov — Method

dcov(A::DistanceMatrix{T}, B::DistanceMatrix{T}) where {T <: Real}

Computes the distance covariance of two doubly-centered distance matrices.

EnergyStatistics.dvar — Method

dvar(A::DistanceMatrix{T}) where {T <: Real}

Computes the distance variance of a doubly-centered distance matrix.

Index

EnergyStatistics.DistanceMatrix

EnergyStatistics.dcenter!

EnergyStatistics.dcor

EnergyStatistics.dcor

EnergyStatistics.dcov

EnergyStatistics.dcov

EnergyStatistics.dvar

EnergyStatistics.dvar

EnergyStatistics.ucenter!